I Finally did it
I successfully added the Quicksort in Quick.java.
I still do not understand it 100% but I understand like 87% of it.
(Thank you StackOverflow^^) So that works. The function to print out the
array at end is a bit of a bodge, but who the fuck cares. Also i deleted
the quickext.java file, as I did not need it.
Coeptus